Add many things in App.js. See description for details.
I first imported all of the created components and used all of them.
I created a class component and passed two things to state.
Then I created an async componentDidMount function that have the fetchedData const which contains the evoked function and it will evoke many times because of componentDidMount then I set the fetchedData const as the value of data state.
Then I created handleCountryChange async function that have one parameter which is country and the country will be passed as a parameter to fetchData function which is imported from ./api and the it have the fetchedData const which contains the evoked function with an await keyword.
Then I passed country from the state to Chart component as props and data also.
Added image on the top.
Passed data to Card component.
Passed handleCountryChange to CountryPicker.
Added credits at the bottom.