Added many things, see description!
Imported necessary things.
Created Cards function and de-structure the props that is passed to it.
Printed the recovered value, deaths and infected value in the UI. Used CountUp. :-)
Printed lastUpdate value used .toDateString and passed the lastUpdate as a parameter to new Date.
Implemented own styles.
Exported the Cards component.