Added many things, see description!
Imported axios for fetching data.
Created a fetchData async arrow function that have the country as a param. It have a let that have the url const.
The changeableUrl's value will be changed if the param is passed /countries/{requestedCountry} will be added to it at the right side.
It have a try, catch. In try, I destructure and get the needed things.
Created fetchDailyData async arrow function and created a try catch and In try I get daily data and destructured it to only get the needed data.
Mapped the data in an object and stored it in a modifiedData const.
Created fetchCountries async arrow function that fetch the name of all countries.
Exported all arrow functions.