Added many things, see description!
Created Chart arrow function and exported it. Created a dailyData and setDailyData as a value for useState and set the default value as []. Set the default font-family of Chart into Fraunces. Used useEffect and Added a fetchAPI async arrow function and in it, I used fetchDailyData that I have exported and set this as a value of dailyData. Used the dailyData in the lineChart. Then I created a barChart and the values are fetched from the props of Chart component. Created ternary operator for both line and bar chart that says that if the data for chart is present then show the chart unless show null. Created ternary operator in the div tag that says if the country is present then show barChart unless show the lineChart. Exported the Chart.