Added many things, see description!
Created a CountryPicker arrow function. Got the handleCountryChange as props from App.js. Created fetchCountries and setFetchCountries for useState and set the default value as [].
Used useEffect hook and added a fetchAPI async arrow function. In it, I used the fetchCountries invoked function and set it as fetchCountries value. Then evoked the fetchAPI inside the useEffect.
Set the value of the onChange of the attribute. In the value I take out the text value and set it as handleCountryChange's param. Printed countries in the options tag.